code reorganization and optimization

+++ Controller/PostController.php
@@ -13,34 +13,17 @@ class PostController
     /**
      * @param $template
      */
-    public function addPost($template)
+    public function addPostForm($template)
     {
         try {
             session_start();
-            if (isset($_POST['title']) && isset($_POST['chapo']) && isset($_POST['user_id'])
-                && isset($_POST['content']) && isset($_POST['published'])) {
-                if (isset($_POST['submit'])) {
-                    $title = strip_tags($_POST['title']);
-                    $chapo = strip_tags($_POST['chapo']);
-                    $user_id = strip_tags($_POST['user_id']);
-                    $content = strip_tags($_POST['content']);
-                    $published = strip_tags($_POST['published']);
-
-                    $token = $_SESSION['token'];
-                    $PostManager = new PostManager();
-                    $PostManager->createPost($title, $chapo, $user_id, $content, $published);
-                    header("Location:index.php?action=dashboard/listPosts&token=$token");
-                } else {
-                    throw new Exception('Tous les champs sont obligatoires');
-                }
-            }
-            if (isset($_SESSION['logged_user']) && isset($_SESSION['token']) && $_SESSION['role'] == 1) {
+            if (isset($_SESSION['logged_user']) && isset($_SESSION['token']) && isset($_SESSION['role']) && $_SESSION['role'] == 1) {
                 if (isset($_GET['token']) && $_GET['token'] == $_SESSION['token']) {
-
                     $logged_user = $_SESSION['logged_user'];
                     $token = $_SESSION['token'];
                     $UserManager = new UserManager();
                     $users = $UserManager->getUsers();
+                    $this->addPost($token);
 
                     echo $template->render(['users' => $users, 'logged_user' => $logged_user, 'token' => $token]);
                 } else {
@@ -51,7 +34,34 @@ class PostController
             }
         } catch (Exception $e) {
             $errorMessage = $e->getMessage();
-            echo $template->render(['errorMessage' => $errorMessage]);
+            echo $template->render(['errorMessage' => $errorMessage, 'users' => $users, 'logged_user' => $logged_user, 'token' => $token]);
+        }
+    }
+
+    /**
+     * @param $token
+     * @throws Exception
+     */
+    private function addPost($token)
+    {
+        if (isset($_POST['title']) && isset($_POST['chapo']) && isset($_POST['user_id']) && isset($_POST['content']) && isset($_POST['published'])) {
+            if (!empty($_POST['title']) && !empty($_POST['chapo']) && !empty($_POST['user_id']) && !empty($_POST['content']) && !empty($_POST['published'])) {
+                if (isset($_POST['submit'])) {
+                    $title = strip_tags($_POST['title']);
+                    $chapo = strip_tags($_POST['chapo']);
+                    $user_id = strip_tags($_POST['user_id']);
+                    $content = strip_tags($_POST['content']);
+                    $published = strip_tags($_POST['published']);
+
+                    $PostManager = new PostManager();
+                    $PostManager->createPost($title, $chapo, $user_id, $content, $published);
+                    header("Location:index.php?action=dashboard/listPosts&token=$token");
+                } else {
+                    throw new Exception('Tous les champs sont obligatoires');
+                }
+            } else {
+                throw new Exception("tous les champs sont obligatoires");
+            }
         }
     }
 
@@ -62,13 +72,14 @@ class PostController
     {
         session_start();
         if (isset($_SESSION['logged_user']) && isset($_SESSION['token']) && isset($_SESSION['role']) && $_SESSION['role'] == 1) {
-            if(isset($_GET['token']) && $_GET['token'] == $_SESSION['token']){
+            if (isset($_GET['token']) && $_GET['token'] == $_SESSION['token']) {
                 $logged_user = $_SESSION['logged_user'];
                 $token = $_SESSION['token'];
                 $postManager = new PostManager();
                 $listPosts = $postManager->getPosts();
+
                 echo $template->render(['listPosts' => $listPosts, 'logged_user' => $logged_user, 'token' => $token]);
-            }else{
+            } else {
                 header("Location:index.php?action=non-existent-backoffice-page");
             }
         } else {
@@ -84,7 +95,7 @@ class PostController
         session_start();
         if (isset($_SESSION['logged_user']) && isset($_SESSION['token']) && isset($_SESSION['role']) && $_SESSION['role'] == 1) {
             $logged_user = $_SESSION['logged_user'];
-            if(isset($_GET['token']) && $_GET['token'] == $_SESSION['token']){
+            if (isset($_GET['token']) && $_GET['token'] == $_SESSION['token']) {
                 $token = $_SESSION['token'];
                 $id = $_GET['id'];
                 $postManager = new PostManager();
@@ -109,13 +120,16 @@ class PostController
     public function editPost($template)
     {
         session_start();
-        if (isset($_SESSION['logged_user']) && isset($_SESSION['token']) && $_SESSION['role'] == 1) {
+        if (isset($_SESSION['logged_user']) && isset($_SESSION['token']) && isset($_SESSION['role']) && $_SESSION['role'] == 1) {
             $logged_user = $_SESSION['logged_user'];
             if (isset($_GET['token']) && $_GET['token'] == $_SESSION['token']) {
                 $token = $_SESSION['token'];
                 $id = $_GET['id'];
+
                 $postManager = new PostManager();
                 $post = $postManager->getPost($id);
+                $this->updatePost($token);
+
                 if ($post) {
                     $UserManager = new UserManager();
                     $users = $UserManager->getUsers();
@@ -132,36 +146,40 @@ class PostController
         }
     }
 
-    public function updatePost()
+    private function updatePost($token)
     {
-        session_start();
-        if (isset($_POST['id']) && isset($_POST['title']) && isset($_POST['chapo']) && isset($_POST['user_id'])
-            && isset($_POST['content']) && isset($_POST['published'])) {
-            $id = $_POST['id'];
-            $token = $_SESSION['token'];
-            $title = strip_tags($_POST['title']);
-            $chapo = strip_tags($_POST['chapo']);
-            $user_id = strip_tags($_POST['user_id']);
-            $content = strip_tags($_POST['content']);
-            $published = $_POST['published'];
+        try {
+            if (isset($_POST['id']) && isset($_POST['title']) && isset($_POST['chapo']) && isset($_POST['user_id']) && isset($_POST['content']) && isset($_POST['published'])) {
+                $id = $_POST['id'];
+                $title = strip_tags($_POST['title']);
+                $chapo = strip_tags($_POST['chapo']);
+                $user_id = strip_tags($_POST['user_id']);
+                $content = strip_tags($_POST['content']);
+                $published = $_POST['published'];
 
-            if (isset($_POST['submit'])) {
-                $postManager = new PostManager();
-                $postManager->updatePost($id, $title, $chapo, $user_id, $content, $published);
-                header("Location:index.php?action=dashboard/listPosts&token=$token");
+                if (isset($_POST['submit'])) {
+                    $postManager = new PostManager();
+                    $postManager->updatePost($id, $title, $chapo, $user_id, $content, $published);
+
+                    header("Location:index.php?action=dashboard/listPosts&token=$token");
+                }
+            } else {
+                throw new Exception("tous les champs sont obligatoires");
             }
+        } catch (Exception $e) {
         }
     }
 
     public function deletePost()
     {
         session_start();
-        if (isset($_SESSION['logged_user']) && $_SESSION['role'] == 1) {
+        if (isset($_SESSION['logged_user']) && isset($_SESSION['role']) && $_SESSION['role'] == 1) {
             if (isset($_GET['token']) && $_GET['token'] == $_SESSION['token']) {
                 $id = $_GET['id'];
                 $token = $_SESSION['token'];
                 $postManager = new PostManager();
                 $post = $postManager->getPost($id);
+
                 if ($post) {
                     $postManager->deletePost($id);
                     header("Location:index.php?action=dashboard/listPosts&token=$token");
@@ -182,7 +200,6 @@ class PostController
     public function post($template)
     {
         session_start();
-
         $id = $_GET['id'];
         $commentManager = new CommentManager();
         $comments = $commentManager->getCommentsPost($id);
@@ -192,19 +209,12 @@ class PostController
         if ($post) {
             if (isset($_SESSION['logged_user'])) {
                 $logged_user = $_SESSION['logged_user'];
-                if(isset($_GET['token']) && $_GET['token'] == $_SESSION['token']){
+                if (isset($_GET['token']) && $_GET['token'] == $_SESSION['token']) {
                     $token = $_SESSION['token'];
-                    if (isset($_POST['content']) && !empty($_POST['content'])) {
-                        $post_id = $id;
-                        $content = strip_tags($_POST['content']);
-                        if (isset($_POST['submit'])) {
-                            $user_id = $_SESSION['id'];
-                            $commentManager->createComment($content, $post_id, $user_id);
-                            header("Location:index.php?action=commentPost&id=$post_id&token=$token");
-                        }
-                    }
+                    $this->commentPost($token, $id, $commentManager);
+
                     echo $template->render(['post' => $post, 'comments' => $comments, 'logged_user' => $logged_user, 'token' => $token]);
-                }else{
+                } else {
                     $token = $_SESSION['token'];
                     header("Location:index.php?action=commentPost&id=$id&token=$token");
                 }
@@ -216,7 +226,25 @@ class PostController
         } else {
             header('Location:index.php?action=non-existent-frontoffice-page');
         }
+    }
 
+    /**
+     * @param $token
+     * @param $id
+     * @param $commentManager
+     */
+    private function commentPost($token, $id, $commentManager)
+    {
+        if (isset($_POST['content']) && !empty($_POST['content'])) {
+            $post_id = $id;
+            $content = strip_tags($_POST['content']);
+            if (isset($_POST['submit'])) {
+                $user_id = $_SESSION['id'];
+                $commentManager->createComment($content, $post_id, $user_id);
+
+                header("Location:index.php?action=commentPost&id=$post_id&token=$token");
+            }
+        }
     }
 
     /**
@@ -231,11 +259,11 @@ class PostController
         if (!isset($_SESSION['logged_user'])) {
             echo $template->render(['listPosts' => $listPosts]);
         } else {
-            if(isset($_SESSION['token'])) {
+            if (isset($_SESSION['token'])) {
                 $token = $_SESSION['token'];
                 $logged_user = $_SESSION['logged_user'];
                 echo $template->render(['listPosts' => $listPosts, 'logged_user' => $logged_user, 'token' => $token]);
-            }else{
+            } else {
                 header('Location:index.php?action=non-existent-frontoffice-page');
             }
         }
